tegra: mce: update echo ari i/o size
MCE SW will echo the 32 bit integer from req.lo to resp.lo and the
value from req.hi to resp.hi. Modify the API to accept 64 bit
integer as input instead of 32 bit integer.

JIRA: TPM-1261

Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2302690
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/platform/tegra/mce/tegra-mce.c
@@ -180,7 +180,7 @@ EXPORT_SYMBOL_GPL(tegra_mce_cc3_ctrl);
  *
  * Returns 0 if success.
  */
-int tegra_mce_echo_data(u32 data, int *matched)
+int tegra_mce_echo_data(u64 data, u64 *matched)
 {
 	if (mce_ops && mce_ops->echo_data)
 		return mce_ops->echo_data(data, matched);
